login required in many functions

+++ website/users.py
@@ -2,33 +2,39 @@ from flask import Blueprint, render_template, redirect, url_for, request, sessio
 from .models import Admin, User, Service, Appointment
 from . import db
 from datetime import datetime
+from flask_login import login_required, current_user
+
 
 
 users = Blueprint('users', __name__)
 
 
 @users.route('/view_users',  methods=['GET', 'POST'])
+@login_required
 def view_users():
-    #!! login required
-    #!! admin required
 
+    if session.get('admin'):
+        user_list = User.query.all()
+        if not user_list:
+            print("users empty")
+        
+        else:
+            print(user_list)
 
-    user_list = User.query.all()
-    if not user_list:
-        print("users empty")
-    
+        return render_template('users.html', user_list = user_list)
+
     else:
-        print(user_list)
-
-    return render_template('users.html', user_list = user_list)
-
+        return ("access to view users denied")
 
 
 @users.route('/create_user', methods = ['GET', 'POST'])
-def create_user():
-    #!! login required
-    #!! admin required
+@login_required
+# maybe login required is problematic due to the import function on signup 
+def create_user(from_signup = None):
 
+    if not session.get('admin') and from_signup == None:
+        print("unauthorized user tried to create a user")
+        return ("unauthorized user tried to create other user")
 
 
     if request.method == "POST":
@@ -37,8 +43,8 @@ def create_user():
 
         email = request.form.get('email')
         password = request.form.get('password')
-
-
+        
+        #!! add a field of "make admin or not" and use an exception to make it works with signup
 
         print(f"Nombre recibido: {name}")
         print(f"Email recibido: {email}")
@@ -52,7 +58,10 @@ def create_user():
         try:
             db.session.add(user)
             db.session.commit()  # Guardar cambios
-            return redirect(url_for("users.view_users"))
+            if not from_signup:
+                return redirect(url_for("users.view_users"))
+            else:
+                return True
             # return jsonify({'success': True, 'message': 'Usuario registrado con éxito',  'redirect_url': url_for('auth.login')}), 201
         
         except Exception as e:
@@ -67,12 +76,8 @@ def create_user():
 
 
 @users.route('/edit_user/<user_id>', methods = ['GET', 'POST'])
+@login_required
 def edit_user(user_id):
-    #!! login required
-
-
-    
-
     if request.method == 'GET':
         return render_template('form_user.html', user_id=user_id,create = 0)
     
@@ -113,21 +118,21 @@ def edit_user(user_id):
 
 
 @users.route('/delete_user/<user_id>')
+@login_required
 def delete_user(user_id):
     print(user_id)
-    #!! login required
-    #!!admin required
+    if session.get('admin'):
+        user = User.query.filter_by(id=user_id).first()
 
+        if user:
+            print("delete", user)
+            db.session.delete(user)
+            db.session.commit()
 
-    user = User.query.filter_by(id=user_id).first()
+        else:
+            return ("error")
 
-    if user:
-        print("delete", user)
-        db.session.delete(user)
-        db.session.commit()
-
+        return redirect(url_for("users.view_users"))
     else:
-        return ("error")
-
-    return redirect(url_for("users.view_users"))
+        return ("access denied to delete user")
